Finishes up the render helper to generate prefilter and irradiance maps.
This is super useful I believe. However, the performance improvement I
was hoping for didn't really seem to be there... It was 190fps before
and now it's 190fps... Oh well.

But hey, now when calling for the irradiance and prefilter maps, you
don't have to do a mix() function every time!