Adds in water bobbing shader.
There are some limitations, i.e. character's butt can clip into the
backside's "water" and the straw in the middle of the bottle also messes
up the illusion where the backside is being used as the top of the
water. May have to get this working in the vertex shader instead.