Adds new render method for water in the bottle. Works well now.
It works by rendering the backside with depth only so it gets written to
the depth buffer and then after that rendering the front side occurs. If
the height of the front side is too high, then a "ray march" to project
the pixel's position to where it'd be if it were on the plane of the
intersection is executed. The resulting depth information is set as the
depth of the fragment and then the depth test in opengl will say whether
or not this should get discarded or not.