Creates new data structure that calculates an AABB bounds for the bottle fill level.
This is much faster and more accurate, since it's two balls we're
calculating for and fewer operations on the cpu. Accuracy can get
increased with a couple more balls, but I'm not really sure if that's
needed. Maybe 4 more for the sides of the bottle, but honestly not
really that much more.

When water is really on the last leg, it gets kinda wonk. There's some
dispute about whether there really is water left and whatnot. Bc of that
dispute, it may be good to go ahead and include those extra balls for
calculation accuracy.