Fixes point lights disappearing at nighttime
Changes the atmospheric scattering transmittance function to be able to
change into a "if there's less light on screen, then that means more
transmittance" in order to solve a problem where point lights were
getting snuffed out by the transmittance being set to 0 bc no light from
the sun was able to wrap around the earth.

Right. Because transmittance in that algorithm means the amount of light
that transmits from the sun to the point on screen. However, if the
one-minus the amount of light actually on screen is greater than this
"transmittance" value, then set the transmittance to that. Technically
incorrect, but I felt like it was a good call. Now the local lights are
able to keep running even at night!