Adds bandaid to clouds taa. It's time to move on.
Not really able to get the prevpostion and currentpos to match up. It
makes sense bc of how the depth buffer really isn't something that can
perfectly show velocity, especially since previous frame occluded things
won't be considered in this velocity buffer. It's pretty unfortunate,
however it's okay. I'm thinking that instead of a TAA approach, in the
future we'll have to do a moving average kind of approach but only for
the noisy parts (the fringes) of the clouds. Man, if we were on the
surface we could've done a botw-style flat cloud system.